Move the OptionTypeMap code closer to home
This is a predecessor to the Configurable PR.  This change moves the OptionTypeInfo maps closer to where they will be used.  When the Configurable changes are adopted, these maps will become static and not associated with the OptionsHelper.

Additionally, the "is_mutable" boolean is changed to an OptionTypeFlags to allow for future expansion/additional use of this field.

Move OptionTypeInfo into its own header file and make it a class

Moved ParseColumnFamilyOption into cf_options, to eliminate duplicate
constant definitions.

The GetStringFromStruct and ParseStructOptions methods were templatized for no benefit.  Internal to these methods, the objects were treated as void pointers.
