+++ db/db_options_test.cc
@@ -38,8 +38,7 @@ class DBOptionsTest : public DBTestBase {
     StringToMap(options_str, &options_map);
     std::unordered_map<std::string, std::string> mutable_map;
     for (const auto opt : db_options_type_info) {
-      if (opt.second.is_mutable &&
-          opt.second.verification != OptionVerificationType::kDeprecated) {
+      if (opt.second.IsMutable() && !opt.second.IsDeprecated()) {
         mutable_map[opt.first] = options_map[opt.first];
       }
     }
@@ -54,8 +53,7 @@ class DBOptionsTest : public DBTestBase {
     StringToMap(options_str, &options_map);
     std::unordered_map<std::string, std::string> mutable_map;
     for (const auto opt : cf_options_type_info) {
-      if (opt.second.is_mutable &&
-          opt.second.verification != OptionVerificationType::kDeprecated) {
+      if (opt.second.IsMutable() && !opt.second.IsDeprecated()) {
         mutable_map[opt.first] = options_map[opt.first];
       }
     }
